requirements.txt
flask==3.0.0
chromadb==0.4.22
sentence-transformers==2.3.1
langchain==0.1.0
langchain-community==0.0.10
openai==1.10.0
python-dotenv==1.0.0
numpy==1.24.3
tiktoken==0.5.2
